refactor(nav): shared scroll and view-transitions in core [PR-02]
Move window scroll and view-transition name helpers into @ecopages/core/client
and delete duplicate implementations from browser-router and react-router.
