+++ packages/browser-router/src/client/services/index.ts
@@ -4,6 +4,5 @@
  */
 
 export { DomSwapper } from './dom-swapper.ts';
-export { ScrollManager } from './scroll-manager.ts';
 export { ViewTransitionManager } from './view-transition-manager.ts';
 export { PrefetchManager } from './prefetch-manager.ts';
