+++ packages/react-router/src/router.ts
@@ -34,8 +34,8 @@ import {
 	shouldInterceptClick,
 } from './navigation.ts';
 import { morphHead } from './head-morpher.ts';
-import { applyViewTransitionNames } from './view-transition-utils.ts';
-import { manageScroll } from './manage-scroll.ts';
+import { applyViewTransitionNames } from '@ecopages/core/client/view-transitions';
+import { manageWindowScroll } from '@ecopages/core/client/scroll';
 import { saveScrollPositions, restoreScrollPositions } from './scroll-persist.ts';
 import {
 	getEcoNavigationRuntime,
@@ -321,7 +321,7 @@ export const EcoRouter: FC<EcoRouterProps> = ({ page, pageProps, options: userOp
 		const previousUrl = new URL(previousUrlRef.current);
 
 		if (url.href !== previousUrl.href) {
-			manageScroll(url, previousUrl, {
+			manageWindowScroll(url, previousUrl, {
 				scrollBehavior: options.scrollBehavior,
 				smoothScroll: options.smoothScroll,
 			});
